Edit Brought Forward & Earning. Logi In Progress

+++ database/migrations/2019_12_06_074342_create_leave_earnings_table.php
@@ -17,9 +17,8 @@ class CreateLeaveEarningsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('leave_type_id')->nullable();
             $table->unsignedBigInteger('user_id')->nullable();
-            $table->integer('no_of_days');
+            $table->float('no_of_days', 8,1);
             $table->timestamps();
-
             $table->foreign('leave_type_id')->references('id')->on('leave_types');
             $table->foreign('user_id')->references('id')->on('users');
